requirements.txt
asgiref==3.5.2
boto3==1.17.58
botocore==1.20.112
cffi==1.15.1
cryptography==38.0.1
Django==4.1.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-storages==1.11.1
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.2.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
